adfimage_write: int types clean-up.

+++ src/adfimage.c
@@ -474,8 +474,9 @@ int adfimage_write ( adfimage_t * const adfimage,
 
     // seek and write the file
     int32_t bytes_written = 0;
-    if ( adfFileSeek ( file, offset ) == RC_OK ) {
-        bytes_written = adfFileWrite ( file, size, ( unsigned char * ) buffer );
+    if ( adfFileSeek ( file, (unsigned) offset ) == RC_OK ) {
+        bytes_written = (int32_t) adfFileWrite ( file, (unsigned) size,
+                                                 ( unsigned char * ) buffer );
     }
 
     // ... and close it
